% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{getMergedData}
\alias{getMergedData}
\title{getMergedData}
\usage{
getMergedData(roundNumber = 1)
}
\arguments{
\item{roundNumber}{An integer to indicate the round from which you are getting the data}
}
\value{
A dataframe with rows converted into their numeric values from the survey codebook. Please work with the codebook
}
\description{
Gets a marged dataset for a given round. The default value is round 1.
}
\examples{
#Extracting a column to check what citizens think about their freedom of expression
merged5 <- getMergedData(roundNumber = 5)
expression_freedom <- merged5 \%>\% dplyr::select(Q17A) \%>\%
 dplyr::mutate(express_freedom = Q17A) \%>\%
 dplyr::select(express_freedom)
str(expression_freedom)

}

