% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Countries.R
\name{getMergedRounds}
\alias{getMergedRounds}
\title{getMergedRounds}
\usage{
getMergedRounds()
}
\value{
A list of all the rounds for which there is a merged dataset available
}
\description{
Gets the all the rounds for which merged datasets are available
}
\examples{

#Basic call
allrounds <- getMergedRounds()
allrounds

}

