% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Countries.R
\name{getNumberofRounds}
\alias{getNumberofRounds}
\title{getNumberofRounds}
\usage{
getNumberofRounds(country)
}
\arguments{
\item{country}{A string character representation of the country whose round data you're looking to get.}
}
\value{
The number of rounds a given country has. Equivalent of \code{length(getRounds(country))}.
}
\description{
Same thing as getRounds, but only returns the number of rounds.
}
\examples{
sen_roundnum <- getNumberofRounds(country='Senegal')
paste('Senegal has ',sen_roundnum, ' rounds of data',sep="",collapse="")

}

