% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{getRoundData}
\alias{getRoundData}
\title{getRoundData}
\usage{
getRoundData(country, round = 1)
}
\arguments{
\item{country}{A string character representation of the country whose round data you're looking to get.}

\item{round}{The given round from which you are grabbing data. The default value is 1.}
}
\value{
A data frame with rows converted into their numeric values from the survey codebook. Please work with the codebook
}
\description{
getRoundData
}
\examples{

sen_r1 <- getRoundData(country='Senegal',round=1)
internet_usage <- sen_r1 \%>\% dplyr::select(Q92B) \%>\%
   dplyr::mutate(CellUsage = Q92B)) \%>\%
   dplyr::select(CellUsage)
str(internet)


}

