% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Countries.R
\name{getRounds}
\alias{getRounds}
\title{getRounds}
\usage{
getRounds(country)
}
\arguments{
\item{country}{A string character representation of the country whose round data you're looking to get.}
}
\value{
The list of all the available rounds (data) your country of selection has.
}
\description{
A helper for figuring out the list of rounds a given country has.
}
\examples{
sen_rounds <- getRounds(country = 'Senegal')
paste("The latest round of Senegal is",gsub('Senegal', '',sen_rounds[1]), sep="",collapse="")

}

